requirements.txt
yfinance>=0.2.36
pandas>=2.0.0
numpy>=1.24.0
ta>=0.10.0  # Technical analysis library
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.0.0
scikit-learn>=1.3.0
beautifulsoup4>=4.12.0
nltk>=3.8.1
transformers>=4.30.0  # For FinBERT
torch>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
tweepy>=4.14.0  # Twitter API client
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0  # Interactive visualizations
streamlit>=1.32.0  # Web interface
